Implemented big refactoring. Created separate files for different tasks
