+++ copy_map.h
@@ -1,0 +1,3 @@
+#include "types.h"
+
+Map* copy_map(Map* m);
