+++ create_map.h
@@ -1,0 +1,4 @@
+#include "types.h"
+
+Map* create_map();
+
