+++ print_map.h
@@ -1,7 +1,4 @@
 #include "types.h"
 
 void render_map(Map* m);
-Map* create_map();
 void print_hero_stats(Map* m);
-Map* copy_map(Map* m);
-
